Changed ClientGroupId in BitfinexNewOrder to Long instead of Integer: https://github.com/jnidzwetzki/bitfinex-v2-wss-api-java/issues/104

+++ src/main/java/com/github/jnidzwetzki/bitfinex/v2/callback/channel/account/info/OrderHandler.java
@@ -92,7 +92,7 @@ public class OrderHandler implements ChannelCallbackHandler {
         order.setOrderId(json.getLong(0));
         final String gid = json.optString(1, null);
         if (gid != null) {
-            order.setClientGroupId(Integer.parseInt(gid));
+            order.setClientGroupId(Long.parseLong(gid));
         }
         final String cid = json.optString(2, null);
         if (cid != null) {
